refactor(canvas): separate shared model state
