+++ apps/canvas-workspace/src/renderer/src/components/chat/ModelCatalog/index.tsx
@@ -2,7 +2,7 @@ import { useEffect, useMemo, useState } from 'react';
 import { useI18n } from '../../../i18n';
 import type { CanvasProviderModel } from '../../../types';
 import { Button, SegmentedControl, TextField } from '../../ui';
-import { matchesModelQuery } from '../modelCatalogUtils';
+import { matchesModelQuery } from '../../../utils/modelCatalog';
 import './index.css';
 
 interface Props {
