+++ apps/canvas-workspace/src/renderer/src/components/chat/ModelSettings.tsx
@@ -2,5 +2,3 @@ import './ModelSettings.css';
 
 export { ModelSwitcher } from './ModelSwitcher';
 export { ModelsSection } from './ModelsSection';
-export { useCanvasModels } from './useCanvasModels';
-export type { ModelSelection, UseCanvasModelsResult } from './modelSettingsTypes';
