+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ModelSwitcher.test.tsx
@@ -4,8 +4,7 @@ import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { I18nProvider } from '../../../i18n';
 import { ModelSwitcher } from '../ModelSwitcher';
-import type { CanvasModelStatus } from '../../../types';
-import type { ModelSelection } from '../modelSettingsTypes';
+import type { CanvasModelStatus, ModelSelection } from '../../../types';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
