+++ apps/canvas-workspace/src/renderer/src/components/chat/hooks/useChatComposerStateKeyed.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react
 import type { AgentContextTabRef, AgentRequestContext, CanvasNode } from '../../../types';
 import type { AgentScope, WorkspaceOption } from '../types';
 import { conversationKeyId, type ConversationKey } from '../../../../../shared/conversation-runtime';
-import { useCanvasModels } from '../ModelSettings';
+import { useCanvasModels } from '../../../hooks/useCanvasModels';
 import { useChatSessions } from './useChatSessions';
 import { useConversationRuntimeStream } from './useConversationRuntimeStream';
 import { useMentions } from './useMentions';
